refactor: split dense gguf parity helpers
