+++ crates/bitnet-cli/src/commands/dense_gguf_linear_parity/mod.rs
@@ -97,25 +97,22 @@ use candle_core::{DType, Device as CandleDevice, IndexOp};
 use clap::Args;
 use memmap2::Mmap;
 use serde_json::{Value, json};
-use sha2::{Digest, Sha256};
 use std::collections::{BTreeMap, BTreeSet};
 use std::fs::File;
 use std::path::{Path, PathBuf};
 
 use crate::planner_receipts::{ExecutionPlanReceiptInput, execution_plan_receipt};
 
+mod digest;
+mod hardware;
+mod roles;
+
+use digest::{sha256_bytes, sha256_f32, sha256_json, sha256_u32, sha256_usize};
+use hardware::{cuda_identity_json, is_rtx5070ti_device_name};
+use roles::{dense_role_label, parse_dense_linear_role, parse_norm_roles, parse_role_sweep};
+
 const HARDWARE_LANE: &str = "nvidia-rtx-5070-ti-cuda";
 const MACHINE_ID: &str = "windows-9950x3d-rtx5070ti";
-const DEFAULT_ROLE_SWEEP: &[DenseGgufTensorRole] = &[
-    DenseGgufTensorRole::AttentionQ,
-    DenseGgufTensorRole::AttentionK,
-    DenseGgufTensorRole::AttentionV,
-    DenseGgufTensorRole::AttentionOutput,
-    DenseGgufTensorRole::MlpGate,
-    DenseGgufTensorRole::MlpUp,
-    DenseGgufTensorRole::MlpDown,
-    DenseGgufTensorRole::Output,
-];
 const DENSE_ONE_LAYER_GAP_CANDIDATE_ORDER: &[&str] =
     &["attention_softmax", "attention_v_mix", "mlp_activation"];
 const DENSE_ONE_LAYER_REMAINING_GAP_CANDIDATE_ORDER: &[&str] = &["mlp_activation"];
@@ -4217,94 +4214,6 @@ fn dense_qwen_d2h_timing_source() -> &'static str {
     "wall_clock_extract_logits_2d_local"
 }
 
-fn parse_dense_linear_role(value: &str) -> Result<DenseGgufTensorRole> {
-    let normalized = value
-        .chars()
-        .filter(|ch| ch.is_ascii_alphanumeric())
-        .collect::<String>()
-        .to_ascii_lowercase();
-    match normalized.as_str() {
-        "output" => Ok(DenseGgufTensorRole::Output),
-        "attentionq" | "attnq" | "q" => Ok(DenseGgufTensorRole::AttentionQ),
-        "attentionk" | "attnk" | "k" => Ok(DenseGgufTensorRole::AttentionK),
-        "attentionv" | "attnv" | "v" => Ok(DenseGgufTensorRole::AttentionV),
-        "attentionoutput" | "attnoutput" | "o" => Ok(DenseGgufTensorRole::AttentionOutput),
-        "mlpgate" | "gate" => Ok(DenseGgufTensorRole::MlpGate),
-        "mlpup" | "up" => Ok(DenseGgufTensorRole::MlpUp),
-        "mlpdown" | "down" => Ok(DenseGgufTensorRole::MlpDown),
-        _ => Err(anyhow!(
-            "unsupported dense linear role `{value}`; expected output, attention_q, attention_k, attention_v, attention_output, mlp_gate, mlp_up, or mlp_down"
-        )),
-    }
-}
-
-fn parse_role_sweep(values: &[String]) -> Result<Vec<DenseGgufTensorRole>> {
-    let roles = if values.is_empty() {
-        DEFAULT_ROLE_SWEEP.to_vec()
-    } else {
-        values.iter().map(|value| parse_dense_linear_role(value)).collect::<Result<Vec<_>>>()?
-    };
-
-    if roles.len() < 2 {
-        bail!("dense GGUF linear role sweep requires at least two roles");
-    }
-
-    let mut seen = BTreeSet::new();
-    for role in &roles {
-        let label = dense_role_label(*role);
-        if !seen.insert(label) {
-            bail!("dense GGUF linear role sweep role `{label}` was requested more than once");
-        }
-    }
-
-    Ok(roles)
-}
-
-fn parse_norm_roles(values: &[String]) -> Result<Vec<DenseGgufTensorRole>> {
-    let roles = if values.is_empty() {
-        vec![DenseGgufTensorRole::AttentionNorm, DenseGgufTensorRole::FfnNorm]
-    } else {
-        values.iter().map(|value| parse_dense_norm_role(value)).collect::<Result<Vec<_>>>()?
-    };
-
-    if roles.len() < 2 {
-        bail!("dense GGUF norm fixture extraction requires attention_norm and ffn_norm roles");
-    }
-
-    let mut seen = BTreeSet::new();
-    for role in &roles {
-        let label = dense_role_label(*role);
-        if !seen.insert(label) {
-            bail!("dense GGUF norm fixture role `{label}` was requested more than once");
-        }
-    }
-    for required in [DenseGgufTensorRole::AttentionNorm, DenseGgufTensorRole::FfnNorm] {
-        if !roles.contains(&required) {
-            bail!(
-                "dense GGUF norm fixture extraction requires role `{}`",
-                dense_role_label(required)
-            );
-        }
-    }
-
-    Ok(roles)
-}
-
-fn parse_dense_norm_role(value: &str) -> Result<DenseGgufTensorRole> {
-    let normalized = value
-        .chars()
-        .filter(|ch| ch.is_ascii_alphanumeric())
-        .collect::<String>()
-        .to_ascii_lowercase();
-    match normalized.as_str() {
-        "attentionnorm" | "attnnorm" | "inputlayernorm" => Ok(DenseGgufTensorRole::AttentionNorm),
-        "ffnnorm" | "postattentionlayernorm" | "postattnnorm" => Ok(DenseGgufTensorRole::FfnNorm),
-        _ => Err(anyhow!(
-            "unsupported dense norm role `{value}`; expected attention_norm or ffn_norm"
-        )),
-    }
-}
-
 fn kernel_fixture_from_extracted(
     fixture: &DenseGgufLinearFixture,
 ) -> Result<DenseGgufLinearGemmFixture> {
@@ -6939,23 +6848,6 @@ fn sanitize_label(value: &str) -> String {
         }
     }
     out.trim_matches('_').to_string()
-}
-
-fn dense_role_label(role: DenseGgufTensorRole) -> &'static str {
-    match role {
-        DenseGgufTensorRole::Output => "output",
-        DenseGgufTensorRole::AttentionQ => "attention_q",
-        DenseGgufTensorRole::AttentionK => "attention_k",
-        DenseGgufTensorRole::AttentionV => "attention_v",
-        DenseGgufTensorRole::AttentionOutput => "attention_output",
-        DenseGgufTensorRole::MlpGate => "mlp_gate",
-        DenseGgufTensorRole::MlpUp => "mlp_up",
-        DenseGgufTensorRole::MlpDown => "mlp_down",
-        DenseGgufTensorRole::TokenEmbedding => "token_embedding",
-        DenseGgufTensorRole::AttentionNorm => "attention_norm",
-        DenseGgufTensorRole::FfnNorm => "ffn_norm",
-        DenseGgufTensorRole::Other => "other",
-    }
 }
 
 fn dense_gguf_norm_fixture_receipt_json(
@@ -13038,90 +12930,8 @@ fn descriptor_element_count(descriptor: &DenseGgufTensorDescriptor) -> usize {
     descriptor.shape.iter().copied().fold(1usize, |acc, dim| acc.saturating_mul(dim)).max(1)
 }
 
-fn cuda_identity_json(probe: Option<&bitnet_device_probe::NvidiaCudaProbe>) -> Value {
-    match probe {
-        Some(probe) => json!({
-            "available": probe.available,
-            "device_count": probe.device_count,
-            "device_index": probe.selected_device_index.unwrap_or(0),
-            "device_name": probe.selected_device_name.clone().unwrap_or_else(|| "unknown".into()),
-            "compute_capability": probe.compute_capability.clone().unwrap_or_else(|| "12.0".into()),
-            "driver_version": probe.driver_version.clone().unwrap_or_else(|| "unknown".into()),
-            "cuda_runtime_version": probe.cuda_runtime_version.clone().unwrap_or_else(|| "unknown".into()),
-            "cuda_toolkit_version": probe.cuda_toolkit_version.clone().unwrap_or_else(|| "unknown".into()),
-            "nvrtc_version": probe.nvrtc_version.clone().unwrap_or_else(|| "unknown".into()),
-            "nvml_available": probe.nvml_available,
-            "vram_bytes": probe.vram_bytes.unwrap_or(1),
-            "power_limit_watts": probe.power_limit_watts,
-            "power_draw_watts": probe.power_draw_watts,
-            "temperature_c": probe.temperature_c,
-        }),
-        None => json!({
-            "available": true,
-            "device_count": 1,
-            "device_index": 0,
-            "device_name": "NVIDIA GeForce RTX 5070 Ti",
-            "compute_capability": "12.0",
-            "driver_version": "591.86",
-            "cuda_runtime_version": "12.9",
-            "cuda_toolkit_version": "12.9",
-            "nvrtc_version": "12.9",
-            "nvml_available": true,
-            "vram_bytes": 17094475776_u64,
-            "power_limit_watts": 300.0,
-            "power_draw_watts": 34.97,
-            "temperature_c": 38.0,
-        }),
-    }
-}
-
-fn sha256_bytes(bytes: &[u8]) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(bytes);
-    format!("{:x}", hasher.finalize())
-}
-
-fn sha256_f32(values: &[f32]) -> String {
-    let mut hasher = Sha256::new();
-    for value in values {
-        hasher.update(value.to_le_bytes());
-    }
-    format!("{:x}", hasher.finalize())
-}
-
-fn sha256_usize(values: &[usize]) -> String {
-    let mut hasher = Sha256::new();
-    for value in values {
-        hasher.update((*value as u64).to_le_bytes());
-    }
-    format!("{:x}", hasher.finalize())
-}
-
-fn sha256_u32(values: &[u32]) -> String {
-    let mut hasher = Sha256::new();
-    for value in values {
-        hasher.update(value.to_le_bytes());
-    }
-    format!("{:x}", hasher.finalize())
-}
-
-fn sha256_json(value: &Value) -> Result<String> {
-    let bytes = serde_json::to_vec(value)?;
-    Ok(sha256_bytes(&bytes))
-}
-
 fn checked_u64_mul(left: u64, right: u64, label: &str) -> Result<u64> {
     left.checked_mul(right).ok_or_else(|| anyhow!("{label} overflowed"))
-}
-
-fn is_rtx5070ti_device_name(name: &str) -> bool {
-    let compact = name
-        .chars()
-        .filter(|ch| ch.is_ascii_alphanumeric())
-        .collect::<String>()
-        .to_ascii_lowercase();
-
-    compact.contains("nvidia") && compact.contains("rtx5070ti")
 }
 
 #[cfg(test)]
